refactor video-only proxy for /v/ endpoint reuse
