+++ proxy/proxy_stream.go
@@ -3,27 +3,18 @@ package proxy
 import (
 	"bufio"
 	"context"
-	"io"
+	"encoding/base64"
+	"fmt"
 	"m3u-stream-merger/utils"
 	"net/http"
 	"net/url"
 	"os"
-	"strconv"
+	"path"
 	"strings"
-	"time"
 )
 
 func (instance *StreamInstance) ProxyStream(ctx context.Context, m3uIndex string, subIndex string, resp *http.Response, r *http.Request, w http.ResponseWriter, statusChan chan int) {
 	debug := os.Getenv("DEBUG") == "true"
-
-	bufferMbInt, err := strconv.Atoi(os.Getenv("BUFFER_MB"))
-	if err != nil || bufferMbInt < 0 {
-		bufferMbInt = 0
-	}
-	buffer := make([]byte, 1024)
-	if bufferMbInt > 0 {
-		buffer = make([]byte, bufferMbInt*1024*1024)
-	}
 
 	if r.Method != http.MethodGet || utils.EOFIsExpected(resp) {
 		scanner := bufio.NewScanner(resp.Body)
@@ -60,7 +51,21 @@ func (instance *StreamInstance) ProxyStream(ctx context.Context, m3uIndex string
 					u = base.ResolveReference(u)
 				}
 
-				_, err = w.Write([]byte(u.String() + "\n"))
+				ext := ""
+				baseSplit := strings.Split(path.Base(r.URL.Path), ".")
+				if len(baseSplit) > 1 {
+					ext = baseSplit[len(baseSplit)-1]
+				}
+
+				origUrl := u.String()
+				encodedUrl := base64.URLEncoding.EncodeToString([]byte(origUrl))
+				newUrl := fmt.Sprintf("%s://%s/v/%s", r.URL.Scheme, r.URL.Host, encodedUrl)
+
+				if ext != "" {
+					newUrl += "." + ext
+				}
+
+				_, err = w.Write([]byte(newUrl + "\n"))
 				if err != nil {
 					utils.SafeLogf("Failed to write URL to response: %v", err)
 					statusChan <- 4
@@ -81,113 +86,5 @@ func (instance *StreamInstance) ProxyStream(ctx context.Context, m3uIndex string
 		instance.Cm.UpdateConcurrency(m3uIndex, false)
 	}()
 
-	defer func() {
-		buffer = nil
-	}()
-
-	timeoutSecond := 3
-	if ts, err := strconv.Atoi(os.Getenv("STREAM_TIMEOUT")); err == nil && ts >= 0 {
-		timeoutSecond = ts
-	}
-
-	timeoutDuration := time.Duration(timeoutSecond) * time.Second
-	if timeoutSecond == 0 {
-		timeoutDuration = time.Minute
-	}
-
-	timeStarted := time.Now()
-	lastErr := timeStarted
-
-	returnStatus := 0
-
-	// Backoff settings
-	initialBackoff := 200 * time.Millisecond
-	maxBackoff := time.Duration(timeoutSecond-1) * time.Second
-	currentBackoff := initialBackoff
-
-	contextSleep := func(ctx context.Context) {
-		select {
-		case <-ctx.Done():
-			return
-		case <-time.After(currentBackoff):
-			currentBackoff *= 2
-			if currentBackoff > maxBackoff {
-				currentBackoff = maxBackoff
-			}
-		}
-	}
-
-	readChan := make(chan struct {
-		n   int
-		err error
-	}, 1)
-
-	for {
-		go func() {
-			n, err := resp.Body.Read(buffer)
-			readChan <- struct {
-				n   int
-				err error
-			}{n, err}
-		}()
-
-		elapsed := time.Since(timeStarted)
-		if timeoutSecond > 0 && elapsed >= timeoutDuration {
-			utils.SafeLogf("Timeout reached while trying to stream: %s\n", r.RemoteAddr)
-			statusChan <- returnStatus
-			return
-		}
-
-		select {
-		case <-ctx.Done():
-			utils.SafeLogf("Context canceled for stream: %s\n", r.RemoteAddr)
-			_ = resp.Body.Close()
-			return
-		case result := <-readChan:
-			switch {
-			case result.err == io.EOF:
-				lastErr = time.Now()
-				if utils.EOFIsExpected(resp) || timeoutSecond == 0 {
-					utils.SafeLogf("Stream ended (expected EOF reached): %s\n", r.RemoteAddr)
-					statusChan <- 2
-					return
-				}
-
-				utils.SafeLogf("Stream ended (unexpected EOF reached): %s\n", r.RemoteAddr)
-				returnStatus = 2
-
-				utils.SafeLogf("Retrying same stream until timeout (%d seconds) is reached...\n", timeoutSecond)
-				contextSleep(ctx)
-			case result.err != nil:
-				lastErr = time.Now()
-				utils.SafeLogf("Error reading stream: %s\n", result.err.Error())
-				returnStatus = 1
-				if timeoutSecond == 0 {
-					statusChan <- 1
-					return
-				}
-
-				utils.SafeLogf("Retrying same stream until timeout (%d seconds) is reached...\n", timeoutSecond)
-				contextSleep(ctx)
-			case result.err == nil:
-				if _, err := w.Write(buffer[:result.n]); err != nil {
-					utils.SafeLogf("Error writing to response: %s\n", err.Error())
-					statusChan <- 0
-					return
-				}
-
-				if flusher, ok := w.(http.Flusher); ok {
-					flusher.Flush()
-				}
-
-				// check if never errored or last error was at least a second ago
-				if lastErr.Equal(timeStarted) || time.Since(lastErr) >= time.Second {
-					// Reset timer on successful read/write
-					timeStarted = time.Now()
-
-					currentBackoff = initialBackoff
-				}
-			}
-		}
-	}
+	statusChan <- ProxyVideoStream(ctx, resp, r, w)
 }
